single page app
Build a React app that uses an API to fetch data and display it on the page. The app should have a loading spinner while data is being fetched and handle error cases. In case there are more than 10 records, a pagination should be available. Sorting and filtering should be also possible.